+++ Src/Libraries/2-Application/Application.Share/Dtos/Workspace/Projects/ProjectReportDto.cs
@@ -8,7 +8,7 @@ namespace TaskoMask.Application.Share.Dtos.Workspace.Projects
    public class ProjectReportDto : BoardReportDto
     {
         [Display(Name = nameof(ApplicationMetadata.BoardsCount), ResourceType = typeof(ApplicationMetadata))]
-        public int BoardsCount { get; set; }
+        public long BoardsCount { get; set; }
 
     }
 }
